文件整理归档

+++ hardware/MIPS/1.0.4/cores/arduino/Serial.h
@@ -1,4 +1,4 @@
-void SetBaudRate(unsigned int buadRate);
+void SetBaudRate(unsigned long buadRate);
 
 void Putch(char ch);
 void Puts(char const* str);
